+++ backend/tests/api/test_defect_detail_put.py
@@ -5,16 +5,17 @@ from tests.factory.user_factory import UserFactory
 from tests.factory.permission_factory import PermissionFactory
 from tests.factory.customer_factory import CustomerFactory
 from tests.factory.order_factory import OrderFactory
+from tests.factory.defect_factory import DefectFactory
 from apps.bug_note.models.defect import Defect
 
 
 @pytest.mark.django_db
-class TestDefectCreateView:
+class TestDefectDetailPut:
     """
-    不具合情報を新規作成するAPIのテスト
+    不具合情報を編集するAPIのテスト
 
-    URL: /api/bug_note/defects/
-    METHOD: POST
+    URL: /api/bug_note/defects/<int:pk>/
+    METHOD: PUT
     """
 
     @pytest.fixture
@@ -32,16 +33,20 @@ class TestDefectCreateView:
         customer = CustomerFactory()
         return OrderFactory(customer=customer)
 
-    def test_create_defect_success(self, client, authed_user, order):
+    @pytest.fixture
+    def defect(self, order, authed_user):
+        return DefectFactory(order=order, create_user=authed_user)
+
+    def test_update_defect_success(self, client, authed_user, order, defect):
         """
-        正常系: 不具合情報を新規作成する
+        正常系: 不具合情報を編集する
 
         条件:
-        - 不具合情報の作成
+        - 不具合情報の編集
 
         結果:
-        - ステータスコード201
-        - 不具合情報が作成される
+        - ステータスコード200
+        - 不具合情報が編集される
         """
         client.force_authenticate(user=authed_user)
         data = {
@@ -53,22 +58,22 @@ class TestDefectCreateView:
             "submission_deadline": "2025-04-10T17:00:00Z"
         }
 
-        response = client.post("/api/bug_note/defects/", data=data, format="json")
+        response = client.put(f"/api/bug_note/defects/{defect.id}/", data=data, format="json")
 
-        assert response.status_code == status.HTTP_201_CREATED
+        assert response.status_code == status.HTTP_200_OK
         assert response.data["title"] == "不具合テスト"
         assert Defect.objects.count() == 1
 
-    def test_unauthenticated_user_cannot_create(self, client, order):
+    def test_unauthenticated_user_cannot_update(self, client, order, defect):
         """
-        異常系: 認証されていないユーザーが作成できない
+        異常系: 認証されていないユーザーが編集できない
 
         条件:
         - 認証されていないユーザー
 
         結果:
         - ステータスコード401
-        - 不具合情報が作成されない
+        - 不具合情報が編集されない
         """
         data = {
             "order": order.id,
@@ -79,19 +84,19 @@ class TestDefectCreateView:
             "submission_deadline": "2025-04-10T17:00:00Z"
         }
 
-        response = client.post("/api/bug_note/defects/", data=data, format="json")
+        response = client.put(f"/api/bug_note/defects/{defect.id}/", data=data, format="json")
         assert response.status_code == status.HTTP_401_UNAUTHORIZED
 
-    def test_user_without_permission_cannot_create(self, client, order):
+    def test_user_without_permission_cannot_update(self, client, order, defect):
         """
-        異常系: 権限がないユーザーが作成できない
+        異常系: 権限がないユーザーが編集できない
 
         条件:
         - 権限がないユーザー
 
         結果:
         - ステータスコード403
-        - 不具合情報が作成されない
+        - 不具合情報が編集されない
         """
         user = UserFactory()
         PermissionFactory(user=user, can_edit_defect=False)
@@ -106,10 +111,10 @@ class TestDefectCreateView:
             "submission_deadline": "2025-04-10T17:00:00Z"
         }
 
-        response = client.post("/api/bug_note/defects/", data=data, format="json")
+        response = client.put(f"/api/bug_note/defects/{defect.id}/", data=data, format="json")
         assert response.status_code == status.HTTP_403_FORBIDDEN
 
-    def test_invalid_data(self, client, authed_user):
+    def test_invalid_data(self, client, authed_user, defect):
         """
         異常系: 不正なデータが送信された場合、エラーが返される
 
@@ -118,8 +123,8 @@ class TestDefectCreateView:
 
         結果:
         - ステータスコード400
-        - 不具合情報が作成されない
+        - 不具合情報が編集されない
         """
         client.force_authenticate(user=authed_user)
-        response = client.post("/api/bug_note/defects/", data={}, format="json")
+        response = client.put(f"/api/bug_note/defects/{defect.id}/", data={}, format="json")
         assert response.status_code == status.HTTP_400_BAD_REQUEST
