+++ backend/tests/api/test_defect_get.py
@@ -9,7 +9,7 @@ from tests.factory.permission_factory import PermissionFactory
 
 
 @pytest.mark.django_db
-class TestDefectListView:
+class TestDefectGet:
     """
     不具合一覧取得APIのテスト
